Creates the BaseStyles Component
