+++ src/components/BaseStyles/index.js
@@ -1,0 +1,6 @@
+// keyframes
+export { float } from './keyframes'
+// typography
+export { SectionTitle, SectionSubTitle, SectionMessage } from './typography'
+// table
+export { Table, Numbered } from './table'
